+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added the English Word in Context dataset [WiC](https://aclanthology.org/N19-1128/),
+  based on the SuperGLUE benchmark. The split is given by 1,024 / 256 / 638 samples for
+  train / val / test, respectively. The train and val splits are stratified subsets of
+  the original SuperGLUE training split, and the test split is the original SuperGLUE
+  validation split. It is marked as `unofficial` for now.
 - Added the Danish metaphor interpretation dataset DAMETA, part of the [Danish Semantic
   Reasoning Benchmark](https://github.com/kuhumcst/danish-semantic-reasoning-benchmark).
   The split is given by 64 / 128 / 723 samples for train / val / test, respectively.